Address potential XSS vulnerability when injecting emoji images
We inject the emoji URL strings returned by the Slack API into the target page's DOM. A
malicious API could return arbitrary HTML in place of the image URL we expect, so we need
to be careful. With this in mind, I previously had added a check to only return URL
strings that match a URL regex.

This commit addresses a bug in the regex: it would match any string that contained a URL,
not just a string that consisted only of a URL. Obviously that's problematic.

As a second line of defence, I also replaced the innerHTML interpolation in the content
script with a more correct node creation approach, which ensures that the only nodes that
get created are the ones we intend to be created. Now, if the URL string somehow contains
arbitrary HTML (which shouldn't be possible with the fixed regex), it will be safely
neutralized by setAttribute.

This change incidentally fixed a bug where escaped HTML entities in the original text
content were unescaped after injecting the emoji. Lesson: don't mess with innerHTML; do
things properly in the first place.

I'll probably make similar changes to options.js, since technically a malicious cookie
from Slack could pwn that file as it's written right now.