Cache missing emoji, too (i.e. don't ask if they exist every time we see them)
This means if someone were to see the text representation of an emoji that didn't exist
then immediatley go and create it, we won't load the emoji until the cache expires or the
user clears it manually. That kind of sucks, but I imagine this situation will beless
common than the case I'm trying to avoid, which is where a non-existent emoji string (or
what looks like an emoji string) is viewed over and over again-- we don't want to ask
Slack about it every time we see it.