Replace options page innerHTML string manipulation with direct DOM element creation
This is much safer because it ensures untrusted data from Slack's localstorage isn't
just parsed as HTML. Technically, we could still use innerHTML and specifically escape
the strings before interpolating them, but it feels less hacky and more in the spirit of
the DOM to do it this way (plus it lessens the risk of future carelessness introducing
a vulnerability).

Follow-up to 263a49b8057370c68e7db521b5a3b5f56f15c4ca.