Move error checking for emoji URLs received from server closer to edge
Also, there's no reason not to just panic if we receive a bad URL from the
server.

Recall that this is part of a 2-line defence against script injection from a
compromised or malicious emoji server; the changes in
263a49b8057370c68e7db521b5a3b5f56f15c4ca mean that even if we were to accept
a malicious string containing, for example, a script element, it wouldn't
have any effect since it would be safely contained in the img element's src
attribute.