Improve emoji picker conflict with enter-to-send behaviour
Workplace by Meta posts an in-progress comment if you press enter. This could cause
a partial emoji name to be posted if you try to press enter to select an emoji from
the picker. Making this event listener capturing and also stopping the event from
propagating prevents an enter keypress from submitting a comment.