Fix bug with inserting end space at wrong position
Note: there's another bug with inserting end spaces. In Chrome, when you type a space
at the end of a contenteditable text node, it actually inserts an nbsp. Then, when you
type another letter, it replaces that nbsp with a regular space. Without the nbsp, the
cursor won't go to the position after the space. Firefox seems to handle trailing
spaces fine. This would be tricky to fix since the method would need to be aware of
both the browser type and the node type (since trailing spaces work fine in textarea/
input nodes). Ok, not that tricky... but messy, which is why I'm going to ignore it
for now.